requirements-dev.txt
Sphinx>=1.6.3
sphinx-rtd-theme>=0.2.4
django-extensions>=1.8.0
coverage>=4.4.1
ipdb>=0.10.3
ipython>=6.1.0
pytest-django>=3.1.2
pytest>=3.1.2
pytest-cov>=2.5.1
tox>=2.7.0
invoke>=0.22
pycodestyle
python3-indy
